pyproject.toml
[project]
name = "image-compression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "streamlit>=1.56.0",
]
